Second iteration

+++ src/main/java/me/riccalioliojr/apps/constants/Constants.java
@@ -1,8 +1,10 @@
 package me.riccalioliojr.apps.constants;
 
+import java.math.BigDecimal;
+
 public class Constants {
-    public static final int DAYS_PER_MONTH = 30;
-    public static final int DAYS_PER_YEAR = 360;
+    public static final BigDecimal DAYS_PER_MONTH = BigDecimal.valueOf(30);
+    public static final BigDecimal DAYS_PER_YEAR = BigDecimal.valueOf(360);
     public static final int DECIMALS = 2;
     public static final int PAYMENTS_PER_YEAR = 12;
 }
